Added miscUtils module
