+++ src/ffpack/utils/__init__.py
@@ -2,3 +2,4 @@ from .generalUtils import *
 from .fdrUtils import *
 from .lccUtils import *
 from .lsmUtils import *
+from .miscUtils import *
